+++ projects/ids-enterprise-ng/src/lib/modal-dialog/soho-modal-dialog.ref.ts
@@ -18,6 +18,12 @@ export class SohoModalDialogRef<T> {
   // tslint:disable-next-line: deprecation
   private eventGuard: SohoModalDialogVetoableEventGuard<T> = {};
 
+  /**
+   * Closes the modal dialogs if router navigation is detected, this prevents diaslogs from being
+   * left open when navigating.
+   */
+  private _closeOnNavigation = true;
+
   /** Selector referencing the modal-dialog after it has been moved to the dialog container. */
   private jQueryElement: JQuery;
 
@@ -237,12 +243,23 @@ export class SohoModalDialogRef<T> {
    * instance.
    *
    * @param component - the instantated instance.
-   * @return the dialof ref for onward assignment.
+   * @return the dialog ref for onward assignment.
    */
   apply(fn: (component: T) => void): SohoModalDialogRef<T> {
     if (fn && this.componentRef.instance) {
       fn(this.componentRef.instance);
     }
+    return this;
+  }
+
+  /**
+   * When set to true, this dialog is closed when navigation is detected.
+   *
+   * @param closeOnNavigation controls the close behaviour when navigating.
+   * @return the dialog ref for support a fluent api.
+   */
+  closeOnNavigation(closeOnNavigation: boolean): SohoModalDialogRef<T> {
+    this._closeOnNavigation = closeOnNavigation;
     return this;
   }
 
@@ -286,21 +303,26 @@ export class SohoModalDialogRef<T> {
         this.close();
       });
 
-      // Add a subscription to the router to remove
-      // the dialog when the user navigates.
-      router.events
-        .pipe(takeUntil(this.destroyed$))
-        .subscribe(e => {
-          if (e instanceof NavigationEnd) {
-            this.componentRef.destroy();
-          }
-        });
-
       // Initialise the event guart
       this.eventGuard = this.componentRef.instance;
 
       this._options.content = jQuery(this.componentRef.location.nativeElement);
     }
+
+    // Add a subscription to the router to remove
+    // the dialog when the user navigates.
+    router.events
+      .pipe(takeUntil(this.destroyed$))
+      .subscribe(e => {
+        if (this._closeOnNavigation && e instanceof NavigationEnd) {
+          if (this.modal) {
+            this.modal.close(true);
+          }
+          if (this.componentRef) {
+            this.componentRef.destroy();
+          }
+        }
+      });
   }
 
   /**
@@ -509,8 +531,6 @@ export class SohoModalDialogRef<T> {
         // Tidy up all subscriptions.
         this.destroyed$.next();
         this.destroyed$.complete();
-
-        console.log(`onBeforeDestroy - published destroy event`);
 
         // Clean up references
         this.eventGuard = null;